Move find grasp_sm to the other components
